pyproject.toml
[project]
name = "biolabs-multiagent"
version = "0.1.0"
description = "Multi-Agent Business Intelligence System for BioLabs"
authors = [
    { name = "Angel Maraver", email = "angel@biolabs.ai" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Databricks SDK
    "databricks-sdk>=0.20.0",
    "databricks-vectorsearch>=0.22",

    # Gradio App
    "gradio>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",

    # MCP Server
    "fastmcp>=0.3.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",

    # Data Generation
    "faker>=20.0.0",
    "python-dateutil>=2.8.0",

    # PDF Processing
    "pypdf>=3.17.0",
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",

    # AI/LLM
    "anthropic>=0.18.0",
    "openai>=1.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
